Finish the 64-bit to int narrowing
The remaining implicit 64-bit to 32-bit conversions sat inside macro
expansions, spanned several lines, or were reported only by GCC.
CuAssertIntEquals arguments and the GET_IDNUM, GET_EXP and GET_PREF
fields are converted explicitly at the use; APPEND_TO_BUF converts its
size_t offset where it calls snprintf_append; experience displays,
string length sums, time fields and MSDP values cast the whole
expression; and float conversions of time and counter values cast the
operand. The clan investment return is now long, since an investment
and its return can exceed int, and its log message uses %ld.

look_at_room_number took a long, so its room_number < 0 guard never
fired for the room_rnum values every caller passes, and NOWHERE reached
ROOM_FLAGS(). It now takes a room_rnum and rejects NOWHERE and rooms
past top_of_world.

Clang 22.1.8 shorten-64-to-32 reaches zero; GCC 16.2 conversion falls
from 792 to 645 sites.

+++ src/act.other.c
@@ -3327,7 +3327,7 @@ void respec_engine(struct char_data *ch, int class, char *arg, bool silent)
   int tempXP;
 
   /* in the clear! */
-  tempXP = GET_EXP(ch);
+  tempXP = (int)GET_EXP(ch);
   original_size = GET_REAL_SIZE(ch);
   preserve_original_size = GET_REAL_RACE(ch) == RACE_LICH || GET_REAL_RACE(ch) == RACE_VAMPIRE;
 
@@ -7170,7 +7170,8 @@ ACMD(do_steal)
       /* Steal some gold coins */
       gold = (GET_GOLD(vict) * rand_number(1, 10)) / 100;
       gold = MIN(1782, gold);
-      gold = long_min(gold, award_capacity(ch, AWARD_GOLD)); /* only what the thief can carry */
+      gold =
+          (int)long_min(gold, award_capacity(ch, AWARD_GOLD)); /* only what the thief can carry */
       if (gold > 0)
       {
         award_gold(ch, gold);
@@ -12033,7 +12034,8 @@ ACMDU(do_device)
     {
       int minutes_left =
           (int)((ch->player_specials->saved.device_creation_cooldown - time(0)) / 60);
-      int seconds_left = (ch->player_specials->saved.device_creation_cooldown - time(0)) % 60;
+      int seconds_left =
+          (int)((ch->player_specials->saved.device_creation_cooldown - time(0)) % 60);
       send_to_char(ch,
                    "You must wait %d minute%s and %d second%s before creating another device.\r\n",
                    minutes_left, (minutes_left == 1) ? "" : "s", seconds_left,
@@ -12582,7 +12584,7 @@ ACMDU(do_device)
     if (inv->cooldown_expires > time(0))
     {
       int hours_left = (int)((inv->cooldown_expires - time(0)) / 3600);
-      int minutes_left = ((inv->cooldown_expires - time(0)) % 3600) / 60;
+      int minutes_left = (int)(((inv->cooldown_expires - time(0)) % 3600) / 60);
       if (hours_left > 0)
       {
         send_to_char(
@@ -13135,7 +13137,7 @@ ACMDU(do_device)
           if (remaining_space > 0)
           {
             strncat(spell_list, spell_name, remaining_space);
-            spell_list_len += strlen(spell_name);
+            spell_list_len += (int)(strlen(spell_name));
             if (spell_list_len >= 195)
             {
               spell_list[195] = '\0';
@@ -13309,7 +13311,7 @@ ACMDU(do_device)
     if (inv->cooldown_expires > time(0) && inv->uses == 0)
     {
       int hours_left = (int)((inv->cooldown_expires - time(0)) / 3600);
-      int minutes_left = ((inv->cooldown_expires - time(0)) % 3600) / 60;
+      int minutes_left = (int)(((inv->cooldown_expires - time(0)) % 3600) / 60);
       send_to_char(ch, "  Status: BROKEN - will be repaired in %d hours, %d minutes\r\n",
                    hours_left, minutes_left);
     }
@@ -13322,7 +13324,7 @@ ACMDU(do_device)
     if (inv->cooldown_expires > time(0))
     {
       int hours_left = (int)((inv->cooldown_expires - time(0)) / 3600);
-      int minutes_left = ((inv->cooldown_expires - time(0)) % 3600) / 60;
+      int minutes_left = (int)(((inv->cooldown_expires - time(0)) % 3600) / 60);
       if (inv->uses == 0)
       {
         send_to_char(ch,
@@ -13613,8 +13615,8 @@ ACMDU(do_device)
       {
         found_any_cooldowns = 1;
         int hours_left = (int)((inv->cooldown_expires - time(0)) / 3600);
-        int minutes_left = ((inv->cooldown_expires - time(0)) % 3600) / 60;
-        int seconds_left = (inv->cooldown_expires - time(0)) % 60;
+        int minutes_left = (int)(((inv->cooldown_expires - time(0)) % 3600) / 60);
+        int seconds_left = (int)((inv->cooldown_expires - time(0)) % 60);
         if (hours_left > 0)
         {
           send_to_char(ch, "  [%d] %s - COOLDOWN: %d hour%s, %d minute%s, %d second%s\r\n", i + 1,
