+++ src/wilderness/wilderness.c
@@ -358,7 +358,7 @@ int get_weather(int x, int y)
   time_t now;
 
   now = time(NULL);
-  time_base = now % 100000;
+  time_base = (double)(now % 100000);
 
   time_base = time_base / (double)(100000.0);
   trans_x = x / (double)(WILD_X_SIZE / 1.0);
